considerable cleanup
