+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/transaction/impl/SnapshotTransactionManager.java
@@ -15,6 +15,7 @@
  */
 package com.palantir.atlasdb.transaction.impl;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
@@ -67,9 +68,9 @@ import com.palantir.lock.v2.LockToken;
 import com.palantir.lock.v2.StartIdentifiedAtlasDbTransactionResponse;
 import com.palantir.lock.v2.TimelockService;
 import com.palantir.logsafe.Preconditions;
+import com.palantir.logsafe.exceptions.SafeRuntimeException;
 import com.palantir.timestamp.TimestampManagementService;
 import com.palantir.timestamp.TimestampService;
-import com.palantir.util.ExceptionHandlingRunner;
 
 /* package */ class SnapshotTransactionManager extends AbstractLockAwareTransactionManager {
     private static final Logger log = LoggerFactory.getLogger(SnapshotTransactionManager.class);
@@ -215,7 +216,7 @@ import com.palantir.util.ExceptionHandlingRunner;
 
     @Override
     public <T, E extends Exception> T finishRunTaskWithLockThrowOnConflict(TransactionAndImmutableTsLock txAndLock,
-            TransactionTask<T, E> task)
+                                                                           TransactionTask<T, E> task)
             throws E, TransactionFailedRetriableException {
         Timer postTaskTimer = getTimer("finishTask");
         Timer.Context postTaskContext;
@@ -299,7 +300,7 @@ import com.palantir.util.ExceptionHandlingRunner;
         }
     }
 
-    private <T, C extends PreCommitCondition, E extends Exception> T runTaskWithConditionReadOnlyInternal(
+    private  <T, C extends PreCommitCondition, E extends Exception> T runTaskWithConditionReadOnlyInternal(
             C condition, ConditionAwareTransactionTask<T, C, E> task) throws E {
         checkOpen();
         long immutableTs = getApproximateImmutableTimestamp();
@@ -345,11 +346,11 @@ import com.palantir.util.ExceptionHandlingRunner;
     /**
      * Frees resources used by this SnapshotTransactionManager, and invokes any callbacks registered to run on close.
      * This includes the cleaner, the key value service (and attendant thread pools), and possibly the lock service.
-     * <p>
-     * Concurrency: If this method races with registerClosingCallback(closingCallback), then closingCallback may be
-     * called (but is not necessarily called). Callbacks registered before the invocation of close() are guaranteed to
-     * be executed (because we use a synchronized list) as long as no exceptions arise. If an exception arises, then no
-     * guarantees are made with regard to subsequent callbacks being executed.
+     *
+     * Concurrency: If this method races with registerClosingCallback(closingCallback), then closingCallback
+     * may be called (but is not necessarily called). Callbacks registered before the invocation of close() are
+     * guaranteed to be executed (because we use a synchronized list) as long as no exceptions arise. If an exception
+     * arises, then no guarantees are made with regard to subsequent callbacks being executed.
      */
     @Override
     public void close() {
@@ -359,19 +360,19 @@ import com.palantir.util.ExceptionHandlingRunner;
             return;
         }
 
-        try (ExceptionHandlingRunner shutdownRunner = new ExceptionHandlingRunner()) {
-            shutdownRunner.runSafely(super::close);
-            shutdownRunner.runSafely(cleaner::close);
-            shutdownRunner.runSafely(keyValueService::close);
-            shutdownRunner.runSafely(() -> shutdownExecutor(deleteExecutor));
-            shutdownRunner.runSafely(() -> shutdownExecutor(getRangesExecutor));
-            shutdownRunner.runSafely(this::closeLockServiceIfPossible);
+        try (ShutdownRunner shutdownRunner = new ShutdownRunner()) {
+            shutdownRunner.shutdownSafely(super::close);
+            shutdownRunner.shutdownSafely(cleaner::close);
+            shutdownRunner.shutdownSafely(keyValueService::close);
+            shutdownRunner.shutdownSafely(() -> shutdownExecutor(deleteExecutor));
+            shutdownRunner.shutdownSafely(() -> shutdownExecutor(getRangesExecutor));
+            shutdownRunner.shutdownSafely(this::closeLockServiceIfPossible);
 
             for (Runnable callback : Lists.reverse(closingCallbacks)) {
-                shutdownRunner.runSafely(callback);
+                shutdownRunner.shutdownSafely(callback);
             }
 
-            shutdownRunner.runSafely(metricsManager::deregisterMetrics);
+            shutdownRunner.shutdownSafely(metricsManager::deregisterMetrics);
             log.info("Close callbacks complete in snapshot transaction manager");
         }
         log.info("Closed snapshot transaction manager without any errors");
@@ -428,8 +429,8 @@ import com.palantir.util.ExceptionHandlingRunner;
     /**
      * This will always return a valid ImmutableTimestamp, but it may be slightly out of date.
      * <p>
-     * This method is used to optimize the perf of read only transactions because getting a new immutableTs requires 2
-     * extra remote calls which we can skip.
+     * This method is used to optimize the perf of read only transactions because getting a new immutableTs requires
+     * 2 extra remote calls which we can skip.
      */
     private long getApproximateImmutableTimestamp() {
         long recentTs = recentImmutableTs.get();
@@ -521,5 +522,27 @@ import com.palantir.util.ExceptionHandlingRunner;
         }
         throw new IllegalArgumentException("Can't use a transaction which is not SnapshotTransaction in "
                 + "SnapshotTransactionManager");
+    }
+
+    private static final class ShutdownRunner implements AutoCloseable {
+        private final List<Throwable> failures = new ArrayList<>();
+
+        void shutdownSafely(Runnable shutdownCallback) {
+            try {
+                shutdownCallback.run();
+            } catch (Throwable throwable) {
+                failures.add(throwable);
+            }
+        }
+
+        @Override
+        public void close() {
+            if (!failures.isEmpty()) {
+                RuntimeException closeFailed = new SafeRuntimeException(
+                        "Close failed. Please inspect the code and fix the failures");
+                failures.forEach(closeFailed::addSuppressed);
+                throw closeFailed;
+            }
+        }
     }
 }
