+++ lock-api-objects/src/main/java/com/palantir/lock/v2/TimelockService.java
@@ -54,8 +54,6 @@ public interface TimelockService {
 
     WaitForLocksResponse waitForLocks(WaitForLocksRequest request);
 
-//    StartIdentifiedAtlasDbTransactionResponseBatch startIdentifiedAtlasDbTransactionBatch(int count);
-
     Set<LockToken> refreshLockLeases(Set<LockToken> tokens);
 
     /**
